Create ondisc_matrix from a single .h5 file or a list of .h5 files.

Creates an ondisc_matrix from a single .h5 file or a list of .h5 files storing single-cell expression data. The .h5 files should fulfill the following format requirements:

There is one and only one dataset named "barcodes";

There is one and only one dataset named "data";

There is one and only one dataset named "gene_names";

There is one and only one dataset named "genes";

There is one and only one dataset named "indices";

There is one and only one dataset named "indptr";

There is one and only one dataset named "shape";

They can be in any h5 group;

Each .h5 file can fit in memory.

create_ondisc_matrix_from_h5_list(
  h5_list,
  odm_fp,
  metadata_fp = NULL,
  barcode_suffixes = NULL,
  progress = TRUE
)

Arguments

h5_list
a single .h5 file or a list of .h5 files to convert into ondisc_matrix format; the .h5 file should have the same features measured on different sets of cells.

odm_fp
location to write the ondisc matrix to disk.

metadata_fp
(optional; default NULL) location to write the metadata .RDS file. By default, a file called "metadata.rds" stored in the same directory as the backing .odm file.

barcode_suffixes
(optional; default NULL) a vector of suffix that appended to each barcodes in input .h5 file(s). The length should be the same as the length of h5_list. If NULL, append nothing for a single .h5 input; append file index for a list of .h5 inputs.

progress
progress (optional; default TRUE) print progress messages?

Value

A covariate_ondisc_matrix.

Details

The function can compute the following cell-specific and feature-specific covariates:

cell-specific: (i) total number of features expressed in cell (n_nonzero_cell), (ii) total UMI count (n_umis_cell), and (iii) percentage of UMIs that map to mitochondrial genes (p_mito_cell).

feature-specific: (i) total number of cells in which feature is expressed (n_nonzero_feature), (ii) mean expression of feature across cells (mean_expression_feature), (iii) coefficient of variation of feature expression across cells (coef_of_variation_feature).

The function decides which covariates to compute given the input; in general, the function computes the maximum set of covariates possible.

Examples

#' # Please install the `ondiscdata` package before running the examples.
# install.packages("devtools")
# devtools::install_github("Katsevich-Lab/ondiscdata")

tempfile <- create_new_directory()
odm_fp <- paste0(tempfile, "/expression_odm")
# get file paths to three .h5 files; these files contain different cells measured on the same genes
h5_list <- system.file("extdata/h5_list", package = "ondiscdata",
c("batch-1_1.h5", "batch-1_2.h5", "batch_2-1.h5"))
# create the ondisc matrix (commented out because of long (~2 min) running time)
# odm_plus_covariates_list <- create_ondisc_matrix_from_h5_list(h5_list, odm_fp)