Create an ondisc_matrix from a single .mtx file or a list of .mtx files.

Initializes an ondisc_matrix from .mtx file(s), a features.tsv file, and barcode.tsv file(s). The number of .mtx files should be the same of the number of the barcodes.tsv files. The list of .mtx files should share the same features.tsv file. Returns an ondisc_matrix along with cell-specific and feature-specific covariate matrices.

create_ondisc_matrix_from_mtx(
  mtx_fp,
  barcodes_fp,
  features_fp,
  odm_fp,
  metadata_fp = NULL,
  n_lines_per_chunk = 3e+08,
  barcode_suffixes = NULL,
  progress = TRUE
)

Arguments

mtx_fp
file path to .mtx file(s) storing the expression data. The .mtx file can represent either an integer matrix or a logical (i.e., binary) matrix. If the .mtx file contains only two columns (after the initial three-column row of metadata), then the .mtx file is assumed to represent a logical matrix.

barcodes_fp
file path to the .tsv file containing the cell barcodes.

features_fp
file path to the features.tsv file. The first column (required) contains the feature IDs (e.g., ENSG00000186092), and the second column (optional) contains the human-readable feature names (e.g., OR4F5). Subsequent columns are discarded.

odm_fp
location to write the ondisc matrix to disk.

metadata_fp
(optional; default NULL) location to write the metadata .RDS file. By default, a file called "metadata.rds" stored in the same directory as the backing .odm file.

n_lines_per_chunk
(optional) number of lines in .mtx file to process per chunk. Defaults to 3e+08.

barcode_suffixes
(optional; default NULL) a vector of suffix that appended to each barcodes in barcodes_fp. The length should be the same as the length of barcodes_fp. If NULL, append nothing for a single .mtx input; append file index for a list of .mtx inputs.

progress
(optional; default TRUE) print progress messages?

Value

A covariate_ondisc_matrix.

Details

The function can compute the following cell-specific and feature-specific covariates:

cell-specific: (i) total number of features expressed in cell (n_nonzero_cell), (ii) total UMI count (n_umis_cell), and (iii) percentage of UMIs that map to mitochondrial genes (p_mito_cell).

feature-specific: (i) total number of cells in which feature is expressed (n_nonzero_feature), (ii) mean expression of feature across cells (mean_expression_feature), (iii) coefficient of variation of feature expression across cells (coef_of_variation_feature).

The function decides which covariates to compute given the input; in general, the function computes the maximum set of covariates possible.

Examples

# # Please install the `ondiscdata` package before running the examples.
# install.packages("devtools")
# devtools::install_github("Katsevich-Lab/ondiscdata")

# First example: initialize a `covariate_ondisc_matrix`
# using simulated expression data; store output in tempdir()
tempfile <- create_new_directory()
odm_fp <- paste0(tempfile, "/expression_odm")
file_locs <- system.file("extdata/mtx", package = "ondiscdata",
c("matrix.mtx", "features.tsv", "barcodes.tsv"))
names(file_locs) <- c("expressions", "features", "barcodes")
expression_data <- create_ondisc_matrix_from_mtx(mtx_fp = file_locs[["expressions"]],
barcodes_fp = file_locs[["barcodes"]],
features_fp = file_locs[["features"]],
n_lines_per_chunk = 537000,
odm_fp = odm_fp)

#> 
#> Writing CSC data.
#> Writing CSR data.
#> 
#> Writing CSC data.
#> Writing CSR data.
#> 
#> Writing CSC data.
#> Writing CSR data.
#> 
#> Writing CSC data.
#> Writing CSR data.

# Second example: initialize from a list of .mtx files
tempfile <- create_new_directory()
odm_fp <- paste0(tempfile, "/expression_odm")
mtx_fp <- system.file("extdata/mtx_list", package = "ondiscdata",
c("mtx_dir1/matrix.mtx", "mtx_dir2/matrix.mtx", "mtx_dir3/matrix.mtx"))
barcodes_fp <- system.file("extdata/mtx_list", package = "ondiscdata",
c("mtx_dir1/barcodes.tsv", "mtx_dir2/barcodes.tsv", "mtx_dir3/barcodes.tsv"))
features_fp <- system.file("extdata", "mtx_list/mtx_dir1/features.tsv", package = "ondiscdata")
odm <- create_ondisc_matrix_from_mtx(mtx_fp, barcodes_fp, features_fp, odm_fp)

#> indexing matrix.mtx [===========--------------------------] 147.58GB/s, eta:  0sindexing matrix.mtx [=====================================] 336.32MB/s, eta:  0s                                                                                indexing matrix.mtx [====---------------------------------] 157.74GB/s, eta:  0sindexing matrix.mtx [=====================================] 269.11MB/s, eta:  0s                                                                                indexing matrix.mtx [======-------------------------------] 229.98GB/s, eta:  0sindexing matrix.mtx [=====================================] 283.15MB/s, eta:  0s                                                                                indexing matrix.mtx [========-----------------------------] 202.64GB/s, eta:  0sindexing matrix.mtx [=====================================] 317.58MB/s, eta:  0s                                                                                
#> Writing CSC data.
#> Writing CSR data.
#> indexing matrix.mtx [====---------------------------------] 145.12GB/s, eta:  0sindexing matrix.mtx [=====================================] 276.65MB/s, eta:  0s                                                                                
#> Writing CSC data.
#> Writing CSR data.
#> indexing matrix.mtx [======-------------------------------] 178.11GB/s, eta:  0sindexing matrix.mtx [=====================================] 279.05MB/s, eta:  0s                                                                                
#> Writing CSC data.
#> Writing CSR data.