Pull a submatrix into memory using the [[ operator.

Apply the [[ operator to an ondisc_matrix or covariate_ondisc_matrix to pull a submatrix into memory. You can pass logical, character, or numeric vectors to [[; character vectors are assumed to refer to feature IDs (for rows) and cell barcodes (for columns).

# S4 method for ondisc_matrix,missing,missing
[[(x, i, j)

# S4 method for ondisc_matrix,ANY,missing
[[(x, i, j)

# S4 method for ondisc_matrix,missing,ANY
[[(x, i, j)

# S4 method for ondisc_matrix,ANY,ANY
[[(x, i, j)

# S4 method for covariate_ondisc_matrix,ANY,ANY
[[(x, i, j)

# S4 method for covariate_ondisc_matrix,missing,ANY
[[(x, i, j)

# S4 method for covariate_ondisc_matrix,ANY,missing
[[(x, i, j)

# S4 method for covariate_ondisc_matrix,missing,missing
[[(x, i, j)

# S4 method for multimodal_ondisc_matrix,ANY,ANY
[[(x, i, j)

Arguments

x
an ondisc_matrix object.

i
a vector (numeric, logical, or character) indicating features to pull into memory.

j
a vector (numeric, logical, or character) indicating cells to pull into memory.

Value

a matrix (as implemented by the Matrix package).

Details

One can remember the difference between [ and [[ by recalling R lists: [ is used to subset a list, and [[ is used to access elements stored inside a list. Similarly, [ is used to subset an ondisc_matrix, and [[ is used to access a submatrix usable within R.

Examples

# Please install the `ondiscdata` package before running the examples.
# install.packages("devtools")
# devtools::install_github("Katsevich-Lab/ondiscdata")

odm_fp <- system.file("extdata", "odm/gene/matrix.odm", package = "ondiscdata")
metadata_fp <- system.file("extdata", "odm/gene/metadata.rds", package = "ondiscdata")
odm <- read_odm(odm_fp, metadata_fp)

# pull a single feature
single_feature <- odm[["ENSG00000180098.9",]]
# pull a single cell
single_cell <- odm[[,"GGAGCAAGTACTTAGC_2_2"]]
# numeric vector
mtx_numeric <- odm[[1:10, 1:10]]
# logical vector
set.seed(1)
nrow_logical <- sample(x = c(TRUE, FALSE), size = nrow(odm), replace = TRUE)
ncol_logical <- sample(x = c(TRUE, FALSE), size = ncol(odm), replace = TRUE)
mtx_logical <- odm[[nrow_logical, ncol_logical]]
# character vectors for feature IDs and cell barcodes
feature_ids_subset <- get_feature_ids(odm)[1:10]
cell_barcodes_subset <- get_cell_barcodes(odm)[1:10]
mtx_char <- odm[[feature_ids_subset, cell_barcodes_subset]]