ondisc_matrix class

An ondisc_matrix represents a feature-by-cell expression matrix stored on-disk.

Construct an ondisc_matrix from an initialized .h5 file.

ondisc_matrix(h5_file)

ondisc_matrix(h5_file)

Arguments

h5_file
a .h5 file storing the on-disk portion of an initialized ondisc_matrix object.

Value

an initialized ondisc_matrix object.

Details

It is best to avoid interacting with the slots of an ondisc_matrix directly. Instead, use the functions and operators provided by the package.

Slots

h5_file

path to an initialized .h5 file stored on-disk.

logical_mat

logical value indicating whether the matrix is logical.

cell_subset

integer vector recording the cells currently in use.

feature_subset

integer vector recording the features currently in use.

underlying_dimension

the dimension of the (unsubset) expression matrix.

Examples

# NOTE: You must create the HDF5 file "expressions.h5" to run this example.
# Navigate to the help file of "create_ondisc_matrix_from_mtx"
# (via ?create_ondisc_matrix_from_mtx), and execute the code in the first code block.
h5_fp <- paste0(tempdir(), "/expressions.h5")
if (file.exists(h5_fp)) {
  odm <- ondisc_matrix(h5_file = h5_fp)
}